requirements.txt
# requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
requests
streamlit
numpy
python-multipart